punto-2-lab-de-digitales/etl/utils.py
import yaml
from pathlib import Path

def load_config(path="config/etl.yaml"):
    with open(path, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)

def ensure_dir(path):
    Path(path).mkdir(parents=True, exist_ok=True)
