punto-2-lab-de-digitales/requirements.txt
pandas==2.2.2
pyyaml==6.0.2
sqlalchemy==2.0.36
numpy==1.26.4
scikit-learn==1.4.2
pyarrow==15.0.2
streamlit==1.39.0
plotly==5.24.0
